# Cache the set of actively monitored blocks