### Resolves

Close #1706 

### Proposed Changes

Cache the set of actively monitored block ids and their corresponding targets.

### Reason for Changes

Projects can have 100's of potentially monitored blocks. Caching the
set of monitored blocks can save time in Runtime._step every frame.

### Numbers

Header | Unit | Description
-- | -- | --
Warm Up, Record | seconds, seconds | Time spent warming up and time spent recording the performance profile
Before | milliseconds | Time spent in Runtime._step per Tick
After | milliseconds | Time spent in Runtime._step per Tick
Change | percent | Percent of difference in duration of time spent outside of Runtime._step per Tick

Project ID | Warm Up, Record | Platform | # of Iterated Monitors | Before | After | Change
-- | -- | -- | -- | -- | -- | --
173918262 | 0s, 5s | 2017 MBP 15" Chrome | 112 | 0.252 | 0.173 | 0.48%
  |   | 2017 MBP 15" Firefox |   | 0.477 | 0.446 | 0.19%
  |   | 2017 MBP 15" Safari |   | 0.450 | 0.358 | 0.55%
  | 5s, 5s | 2017 MBP 15" Chrome |   | 0.250 | 0.155 | 0.57%
  |   | 2017 MBP 15" Firefox |   | 0.432 | 0.426 | 0.04%
  |   | 2017 MBP 15" Safari |   | 0.483 | 0.345 | 0.83%
155128646 | 0s, 5s | 2017 MBP 15" Chrome | 60 | 0.144 | 0.124 | 0.12%
  |   | 2017 MBP 15" Firefox |   | 0.379 | 0.263 | 0.69%
  |   | 2017 MBP 15" Safari |   | 0.360 | 0.157 | 1.21%
  | 5s, 5s | 2017 MBP 15" Chrome |   | 0.155 | 0.114 | 0.25%
  |   | 2017 MBP 15" Firefox |   | 0.316 | 0.329 | -0.08%
  |   | 2017 MBP 15" Safari |   | 0.233 | 0.224 | 0.06%
89811578 | 0s, 5s | 2017 MBP 15" Chrome | 231 | 0.602 | 0.456 | 0.88%
  |   | 2017 MBP 15" Firefox |   | 1.352 | 0.936 | 2.49%
  |   | 2017 MBP 15" Safari |   | 1.034 | 0.785 | 1.49%
  | 5s, 5s | 2017 MBP 15" Chrome |   | 0.465 | 0.365 | 0.60%
  |   | 2017 MBP 15" Firefox |   | 1.329 | 1.100 | 1.38%
  |   | 2017 MBP 15" Safari |   | 1.163 | 0.962 | 1.20%