@joker314 2018-11-07
Those are flaky tests that I've seen fail many times, but closing and reopening (forcing the tests to re-run) will usually make them pass